+++ resources/views/Consignee/Joborders/styles.blade.php
@@ -1,7 +1,5 @@
     {{-- <link href={{URL::asset('/others/stisla_admin_v1.0.0/dist/modules/tempus-dominus/css/tempusdominus-date.css')}} rel="stylesheet" media="all"> --}}
     <link href={{URL::asset('/others/stisla_admin_v1.0.0/dist/modules/datatables/datatables.min.css')}} rel="stylesheet" media="all">
     <link href={{URL::asset('/others/stisla_admin_v1.0.0/dist/css/sweetalert.css')}} rel="stylesheet" media="all">
-    <link href={{URL::asset('/others/stisla_admin_v1.0.0/dist/modules/waves/waves.css')}} rel="stylesheet" media="all">
-    <link href={{URL::asset('/others/stisla_admin_v1.0.0/dist/css/animate.min.css')}} rel="stylesheet" media="all">
     <link href={{URL::asset('/others/stisla_admin_v1.0.0/dist/css/cssAll.css')}} rel="stylesheet" media="all">
-    <link href="/others/stisla_admin_v1.0.0/dist/modules/tempus-dominus/css/tempusdominus-date.css" rel="stylesheet">
+    <link href="/others/ConsigneeAssets/css/ConsigneeJobOrder.css" rel="stylesheet" media="all">