+++ resources/views/ConsigneeAccounts/styles.blade.php
@@ -1,1 +1,2 @@
     <link rel="stylesheet" href="/others/stisla_admin_v1.0.0/dist/css/consigneeAccounts/consigneeAccounts.css">
+